remove active_job, active_record, and action_mailer from required frameworks
since this is supposed to be a slimmed down application focused on just indexing things from fedora to solr, and we want to be able to spin instances up easily to meet high indexing loads, it seems to make sense to not include things unnecessarily.

i imagine we won't be using the following in this app in the near future:
* active_job: i could see adding this back if we decide to have dor_indexing_app handle the bulk indexing that argo does now, though that's slightly trickier to extricate, and really what we want to do there is slim down the size of the workers that do the indexing, which doesn't necessarily mean moving it outside of argo.
* active_record: no need to do any database stuff at this point.  i get the impression that no session management is needed here, since authentication between dlss services seems to happen via IP whitelisting and/or certificate-based auth.
* action_mailer: no plans to send email.

there are notes as to what corresponding settings were commented out, since they would only be relevant to the things that were removed, and caused errors when trying to start the application without the frameworks to which they pertained.