[SPIKE] Index virtual objects as both items and virtual objects
This would make it easier for Argo users to find and operate in bulk on virtual objects. Whether it is valuable, or has other risks or side effects, would be worth discussing. I could also be convinced that we could/should use the existing has_constituents_ssim field for this somehow?