Index all present catalog link types into `metadata_source_ssim`
Fixes #946

This commit adds new indexing code to begin populating a new facet field allowing metadata sources to be multi-valued. This work is being done to support the migration from Symphony to Folio, during which objects will be expected to have one or two metadata sources. It is backwards compatible with the prior behavior---the commit is purely additive.