Rename specs to match file path
RSpec/FilePath was complaining about our spec files not matching the
module namespace. This commit pushes the specs into the directory that
RSpec/FilePath and RSpec/SpecFilePathFormat are expecting. It's not
exactly clear what is going on with the rules, but this seems like the
easiest way to not continue dancing around with the errors is to move
them.

See: https://github.com/rubocop/rubocop-rspec/issues/1717